pyproject.toml
[project]
name = "monitor-manual"
version = "0.1.0"
description = ""
authors = [
    {name = "Felipeflowers17",email = "143847123+Felipeflowers17@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "pandas (>=3.0.0,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pyside6 (>=6.10.1,<7.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "playwright (>=1.57.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"